Move StackAnalysis to EarlyFunctionAnalysis
