+++ include/revng/EarlyFunctionAnalysis/PromoteGlobalToLocalVars.h
