+++ lib/CMakeLists.txt
@@ -23,7 +23,7 @@ add_subdirectory(DwarfImporter)
 add_subdirectory(FunctionCallIdentification)
 add_subdirectory(FunctionIsolation)
 add_subdirectory(Model)
-add_subdirectory(StackAnalysis)
+add_subdirectory(EarlyFunctionAnalysis)
 add_subdirectory(Support)
 add_subdirectory(UnitTestHelpers)
 add_subdirectory(TypeShrinking)
