+++ lib/EarlyFunctionAnalysis/ABIAnalyses/MFIGraphs/ABIAnalysis.template
@@ -15,7 +15,7 @@
 #include "revng/BasicAnalyses/GeneratedCodeBasicInfo.h"
 #include "revng/MFP/MFP.h"
 #include "revng/Model/Binary.h"
-#include "revng/StackAnalysis/StackAnalysis.h"
+#include "revng/EarlyFunctionAnalysis/EarlyFunctionAnalysis.h"
 #include "revng/Support/revng.h"
 
 namespace ABIAnalyses::%LatticeName% {
